Interface segregation and removing unecessary loops
