+++ Interview/Interfaces/IReader.cs
@@ -1,0 +1,8 @@
+﻿namespace Interview.Interfaces
+{
+    public interface IReader
+    {
+        bool Read();
+
+    }
+}
