+++ Interview/Interfaces/IWriter.cs
@@ -1,0 +1,7 @@
+﻿namespace Interview.Interfaces
+{
+    public interface IWriter
+    {
+        void Write(params string[] columns);
+    }
+}
